:recycle: chunking dataclass 분리
